pyproject.toml
[project]
name = "flight-pipeline-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==2.6.3",
    "dbt-core>=1.10.8",
    "dbt-snowflake==1.9.2",
    "google-cloud-storage>=3.3.0",
    "ipykernel>=6.30.1",
    "pandas>=2.3.1",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "snowflake-connector-python>=3.16.0",
    "snowflake-sqlalchemy>=1.7.6",
]
